refactor: model 패키지 정리
