+++ app/src/main/java/woowacourse/movie/domain/date/scheduler/ScreeningDayType.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.domain.date.scheduler
 
 import java.time.DayOfWeek
 import java.time.LocalDate
