+++ app/src/main/java/woowacourse/movie/domain/model/movie/Movie.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.domain.model.movie
 
 import java.io.Serializable
