+++ app/src/main/java/woowacourse/movie/domain/model/movie/RunningTime.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.domain.model.movie
 
 @JvmInline
 value class RunningTime(val time: Int) {
