+++ app/src/main/java/woowacourse/movie/domain/model/reservation/Price.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.domain.model.reservation
 
 @JvmInline
 value class Price(val value: Int) {
