+++ app/src/main/java/woowacourse/movie/domain/model/reservation/PurchaseType.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.domain.model.reservation
 
 import java.time.LocalTime
