+++ app/src/main/java/woowacourse/movie/ui/main/adapter/MoviesAdapter.kt
@@ -4,7 +4,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
-import woowacourse.movie.domain.model.Movie
+import woowacourse.movie.domain.model.movie.Movie
 
 class MoviesAdapter(
     private val movies: List<Movie>,
