+++ app/src/main/java/woowacourse/movie/ui/reserve/ReservePresenter.kt
@@ -3,9 +3,9 @@ package woowacourse.movie.ui.reserve
 import woowacourse.movie.domain.date.scheduler.CurrentDateScheduler
 import woowacourse.movie.domain.date.scheduler.CurrentTimeScheduler
 import woowacourse.movie.domain.date.scheduler.ReservationScheduler
-import woowacourse.movie.domain.model.Movie
-import woowacourse.movie.domain.model.PurchaseCount
-import woowacourse.movie.domain.model.Reservation
+import woowacourse.movie.domain.model.movie.Movie
+import woowacourse.movie.domain.model.reservation.PurchaseCount
+import woowacourse.movie.domain.model.reservation.Reservation
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.LocalTime
